Namespace Session tests
