+++ src/Joomla/Session/Tests/Storage/ApcTest.php
@@ -4,19 +4,23 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
-use Joomla\Session\Storage\Wincache as StorageWincache;
+namespace Joomla\Session\Tests\Storage;
+
+use Joomla\Session\Storage\Apc as StorageApc;
 use Joomla\Session\Storage;
 
 /**
- * Test class for JSessionStorageWincache.
- * Generated by PHPUnit on 2011-10-26 at 19:36:01.
+ * Test class for Joomla\Session\Storage\Apc.
  *
  * @since  1.0
  */
-class JSessionStorageWincacheTest extends PHPUnit_Framework_TestCase
+class ApcTest extends \PHPUnit_Framework_TestCase
 {
 	/**
-	 * @var JSessionStorageWincache
+	 * Test object
+	 *
+	 * @var    StorageApc
+	 * @since  1.0
 	 */
 	protected $object;
 
@@ -24,17 +28,21 @@ class JSessionStorageWincacheTest extends PHPUnit_Framework_TestCase
 	 * Sets up the fixture, for example, opens a network connection.
 	 * This method is called before a test is executed.
 	 *
-	 * @return void
+	 * @return  void
+	 *
+	 * @since   1.0
 	 */
 	protected function setUp()
 	{
-		// Skip these tests if WinCache isn't available.
-		if (!StorageWincache::isSupported())
+		parent::setUp();
+
+		// Skip these tests if APC isn't available.
+		if (!StorageAPC::isSupported())
 		{
-			$this->markTestSkipped('WinCache storage is not enabled on this system.');
+			$this->markTestSkipped('APC storage is not enabled on this system.');
 		}
 
-		$this->object = Storage::getInstance('Wincache');
+		$this->object = Storage::getInstance('APC');
 	}
 
 	/**
