+++ src/Joomla/Session/Tests/Storage/DatabaseTest.php
@@ -1,0 +1,102 @@
+<?php
+/**
+ * @copyright  Copyright (C) 2005 - 2013 Open Source Matters, Inc. All rights reserved.
+ * @license    GNU General Public License version 2 or later; see LICENSE
+ */
+
+namespace Joomla\Session\Tests\Storage;
+
+use Joomla\Session\Storage\Database as StorageDatabase;
+use Joomla\Session\Storage;
+use Joomla\Test\TestDatabase;
+
+/**
+ * Test class for Joomla\Session\Storage\Database.
+ *
+ * @since  1.0
+ */
+class DatabaseTest extends TestDatabase
+{
+	/**
+	 * Test object
+	 *
+	 * @var    StorageDatabase
+	 * @since  1.0
+	 */
+	protected $object;
+
+	/**
+	 * Sets up the fixture, for example, opens a network connection.
+	 * This method is called before a test is executed.
+	 *
+	 * @return  void
+	 *
+	 * @since   1.0
+	 */
+	protected function setUp()
+	{
+		parent::setUp();
+
+		$this->object = Storage::getInstance('Database', array('db' => static::$driver));
+	}
+
+	/**
+	 * Test...
+	 *
+	 * @todo Implement testRead().
+	 *
+	 * @return void
+	 */
+	public function testRead()
+	{
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
+	}
+
+	/**
+	 * Test...
+	 *
+	 * @todo Implement testWrite().
+	 *
+	 * @return void
+	 */
+	public function testWrite()
+	{
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
+	}
+
+	/**
+	 * Test...
+	 *
+	 * @todo Implement testDestroy().
+	 *
+	 * @return void
+	 */
+	public function testDestroy()
+	{
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
+	}
+
+	/**
+	 * Test...
+	 *
+	 * @todo Implement testGc().
+	 *
+	 * @return void
+	 */
+	public function testGc()
+	{
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
+	}
+}
