+++ src/Joomla/Session/Tests/Storage/MemcacheTest.php
@@ -4,47 +4,45 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
+namespace Joomla\Session\Tests\Storage;
+
+use Joomla\Session\Storage\Memcache as StorageMemcache;
+use Joomla\Session\Storage;
+
 /**
- * Test class for JSessionStorage.
- * Generated by PHPUnit on 2011-10-26 at 19:33:06.
+ * Test class for Joomla\Session\Storage\Memcache.
  *
  * @since  1.0
  */
-class JSessionStorageTest extends PHPUnit_Framework_TestCase
+class MemcacheTest extends \PHPUnit_Framework_TestCase
 {
 	/**
-	 * @var JSessionStorage
+	 * Test object
+	 *
+	 * @var    StorageMemcache
+	 * @since  1.0
 	 */
 	protected $object;
 
 	/**
-	 * Test getInstance
+	 * Sets up the fixture, for example, opens a network connection.
+	 * This method is called before a test is executed.
 	 *
-	 * @todo Implement testGetInstance().
+	 * @return  void
 	 *
-	 * @return void
+	 * @since   1.0
 	 */
-	public function testGetInstance()
+	protected function setUp()
 	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
-	}
+		parent::setUp();
 
-	/**
-	 * Test...
-	 *
-	 * @todo Implement testRegister().
-	 *
-	 * @return void
-	 */
-	public function testRegister()
-	{
-		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		// Skip these tests if Memcache isn't available.
+		if (!StorageMemcache::isSupported())
+		{
+			$this->markTestSkipped('Memcache storage is not enabled on this system.');
+		}
+
+		$this->object = Storage::getInstance('Memcache');
 	}
 
 	/**
@@ -57,9 +55,7 @@ class JSessionStorageTest extends PHPUnit_Framework_TestCase
 	public function testOpen()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
@@ -72,9 +68,7 @@ class JSessionStorageTest extends PHPUnit_Framework_TestCase
 	public function testClose()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
@@ -87,9 +81,7 @@ class JSessionStorageTest extends PHPUnit_Framework_TestCase
 	public function testRead()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
@@ -102,9 +94,7 @@ class JSessionStorageTest extends PHPUnit_Framework_TestCase
 	public function testWrite()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
@@ -117,9 +107,7 @@ class JSessionStorageTest extends PHPUnit_Framework_TestCase
 	public function testDestroy()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
@@ -132,9 +120,7 @@ class JSessionStorageTest extends PHPUnit_Framework_TestCase
 	public function testGc()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
@@ -147,8 +133,6 @@ class JSessionStorageTest extends PHPUnit_Framework_TestCase
 	public function testIsSupported()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 }
