+++ src/Joomla/Session/Tests/Storage/NoneTest.php
@@ -4,18 +4,23 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
+namespace Joomla\Session\Tests\Storage;
+
+use Joomla\Session\Storage\None as StorageNone;
 use Joomla\Session\Storage;
 
 /**
- * Test class for JSessionStorageNone.
- * Generated by PHPUnit on 2011-10-26 at 19:36:08.
+ * Test class for Joomla\Session\Storage\None.
  *
  * @since  1.0
  */
-class JSessionStorageNoneTest extends PHPUnit_Framework_TestCase
+class NoneTest extends \PHPUnit_Framework_TestCase
 {
 	/**
-	 * @var JSessionStorageNone
+	 * Test object
+	 *
+	 * @var    StorageNone
+	 * @since  1.0
 	 */
 	protected $object;
 
@@ -23,7 +28,9 @@ class JSessionStorageNoneTest extends PHPUnit_Framework_TestCase
 	 * Sets up the fixture, for example, opens a network connection.
 	 * This method is called before a test is executed.
 	 *
-	 * @return void
+	 * @return  void
+	 *
+	 * @since   1.0
 	 */
 	protected function setUp()
 	{
@@ -33,9 +40,11 @@ class JSessionStorageNoneTest extends PHPUnit_Framework_TestCase
 	}
 
 	/**
-	 * Test JSessionStorageNone::Register().
+	 * Test Joomla\Session\Storage\None::register().
 	 *
-	 * @return void
+	 * @return  void
+	 *
+	 * @since   1.0
 	 */
 	public function testRegister()
 	{
