+++ src/Joomla/Session/Tests/Storage/WincacheTest.php
@@ -4,19 +4,23 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
-use Joomla\Session\Storage\Xcache as StorageXcache;
+namespace Joomla\Session\Tests\Storage;
+
+use Joomla\Session\Storage\Wincache as StorageWincache;
 use Joomla\Session\Storage;
 
 /**
- * Test class for JSessionStorageXcache.
- * Generated by PHPUnit on 2011-10-26 at 19:36:09.
+ * Test class for Joomla\Session\Storage\Wincache.
  *
  * @since  1.0
  */
-class JSessionStorageXcacheTest extends PHPUnit_Framework_TestCase
+class WincacheTest extends \PHPUnit_Framework_TestCase
 {
 	/**
-	 * @var JSessionStorageXcache
+	 * Test object
+	 *
+	 * @var    StorageWincache
+	 * @since  1.0
 	 */
 	protected $object;
 
@@ -24,19 +28,19 @@ class JSessionStorageXcacheTest extends PHPUnit_Framework_TestCase
 	 * Sets up the fixture, for example, opens a network connection.
 	 * This method is called before a test is executed.
 	 *
-	 * @return void
+	 * @return  void
+	 *
+	 * @since   1.0
 	 */
 	protected function setUp()
 	{
-		parent::setUp();
-
-		// Skip these tests if XCache isn't available.
-		if (!StorageXcache::isSupported())
+		// Skip these tests if WinCache isn't available.
+		if (!StorageWincache::isSupported())
 		{
-			$this->markTestSkipped('XCache storage is not enabled on this system.');
+			$this->markTestSkipped('WinCache storage is not enabled on this system.');
 		}
 
-		$this->object = Storage::getInstance('Xcache');
+		$this->object = Storage::getInstance('Wincache');
 	}
 
 	/**
