+++ src/Joomla/Session/Tests/Storage/XcacheTest.php
@@ -4,20 +4,23 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
-use Joomla\Session\Storage\Database as StorageDatabase;
+namespace Joomla\Session\Tests\Storage;
+
+use Joomla\Session\Storage\Xcache as StorageXcache;
 use Joomla\Session\Storage;
-use Joomla\Test\TestDatabase;
 
 /**
- * Test class for JSessionStorageDatabase.
- * Generated by PHPUnit on 2011-10-26 at 19:36:07.
+ * Test class for Joomla\Session\Storage\Xcache.
  *
  * @since  1.0
  */
-class JSessionStorageDatabaseTest extends TestDatabase
+class XcacheTest extends \PHPUnit_Framework_TestCase
 {
 	/**
-	 * @var StorageDatabase
+	 * Test object
+	 *
+	 * @var    StorageXcache
+	 * @since  1.0
 	 */
 	protected $object;
 
@@ -31,9 +34,13 @@ class JSessionStorageDatabaseTest extends TestDatabase
 	{
 		parent::setUp();
 
-		$this->object = Storage::getInstance('Database', array(
-				'db' => self::$driver
-			));
+		// Skip these tests if XCache isn't available.
+		if (!StorageXcache::isSupported())
+		{
+			$this->markTestSkipped('XCache storage is not enabled on this system.');
+		}
+
+		$this->object = Storage::getInstance('Xcache');
 	}
 
 	/**
@@ -46,9 +53,7 @@ class JSessionStorageDatabaseTest extends TestDatabase
 	public function testRead()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
@@ -61,9 +66,7 @@ class JSessionStorageDatabaseTest extends TestDatabase
 	public function testWrite()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
@@ -76,23 +79,19 @@ class JSessionStorageDatabaseTest extends TestDatabase
 	public function testDestroy()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 
 	/**
 	 * Test...
 	 *
-	 * @todo Implement testGc().
+	 * @todo Implement testIsSupported().
 	 *
 	 * @return void
 	 */
-	public function testGc()
+	public function testIsSupported()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete(
-			'This test has not been implemented yet.'
-		);
+		$this->markTestIncomplete('This test has not been implemented yet.');
 	}
 }
