+++ src/Joomla/Session/Tests/StorageTest.php
@@ -4,39 +4,49 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
-use Joomla\Session\Storage\Memcache as StorageMemcache;
-use Joomla\Session\Storage;
+namespace Joomla\Session\Tests;
 
 /**
- * Test class for JSessionStorageMemcache.
- * Generated by PHPUnit on 2011-10-26 at 19:36:04.
+ * Test class for Joomla\Session\Storage.
  *
  * @since  1.0
  */
-class JSessionStorageMemcacheTest extends PHPUnit_Framework_TestCase
+class StorageTest extends \PHPUnit_Framework_TestCase
 {
 	/**
-	 * @var JSessionStorageMemcache
+	 * @var    \Joomla\Session\Storage
+	 * @since  1.0
 	 */
 	protected $object;
 
 	/**
-	 * Sets up the fixture, for example, opens a network connection.
-	 * This method is called before a test is executed.
+	 * Test getInstance
+	 *
+	 * @todo Implement testGetInstance().
 	 *
 	 * @return void
 	 */
-	protected function setUp()
+	public function testGetInstance()
 	{
-		parent::setUp();
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
+	}
 
-		// Skip these tests if Memcache isn't available.
-		if (!StorageMemcache::isSupported())
-		{
-			$this->markTestSkipped('Memcache storage is not enabled on this system.');
-		}
-
-		$this->object = Storage::getInstance('Memcache');
+	/**
+	 * Test...
+	 *
+	 * @todo Implement testRegister().
+	 *
+	 * @return void
+	 */
+	public function testRegister()
+	{
+		// Remove the following lines when you implement this test.
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
 	}
 
 	/**
@@ -49,7 +59,9 @@ class JSessionStorageMemcacheTest extends PHPUnit_Framework_TestCase
 	public function testOpen()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
 	}
 
 	/**
@@ -62,7 +74,9 @@ class JSessionStorageMemcacheTest extends PHPUnit_Framework_TestCase
 	public function testClose()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
 	}
 
 	/**
@@ -75,7 +89,9 @@ class JSessionStorageMemcacheTest extends PHPUnit_Framework_TestCase
 	public function testRead()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
 	}
 
 	/**
@@ -88,7 +104,9 @@ class JSessionStorageMemcacheTest extends PHPUnit_Framework_TestCase
 	public function testWrite()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
 	}
 
 	/**
@@ -101,7 +119,9 @@ class JSessionStorageMemcacheTest extends PHPUnit_Framework_TestCase
 	public function testDestroy()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
 	}
 
 	/**
@@ -114,7 +134,9 @@ class JSessionStorageMemcacheTest extends PHPUnit_Framework_TestCase
 	public function testGc()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
 	}
 
 	/**
@@ -127,6 +149,8 @@ class JSessionStorageMemcacheTest extends PHPUnit_Framework_TestCase
 	public function testIsSupported()
 	{
 		// Remove the following lines when you implement this test.
-		$this->markTestIncomplete('This test has not been implemented yet.');
+		$this->markTestIncomplete(
+			'This test has not been implemented yet.'
+		);
 	}
 }
